Fix comment follow make_list_category
